Add api_handlers_reference.go. Now api_handlers.go is generating automatically (empty right now)
